refactor(api): split flat api client into per-domain namespaces [SHR-188]
Split the single ~736-line src/lib/api.ts into per-domain namespaces over one
shared request core:

- src/lib/api/client.ts — BASE + in-flight GET dedup + request/doRequest
- src/lib/api/taskApi.ts — tasks, diffs, comments, agents, worktrees, chats,
  browse/branches/preflight
- src/lib/api/reviewApi.ts — review inbox/detail/publish + learnings
- src/lib/api/configApi.ts — settings, setup, skills, agents, repo configs,
  integrations, hook registry
- src/lib/api/index.ts — reassembles the flat `api` object, re-exports all
  namespace types + the orchestrator client

src/lib/api.ts is now a thin re-export shim so every existing @/lib/api import
resolves unchanged. Pure refactor; no consumer edits, no behavior change.
